[WIP] Implement ConvolutionCyclopeptideSequencing
Code Challenge: Implement ConvolutionCyclopeptideSequencing.
Input: An integer M, an integer N, and a collection of (possibly
repeated) integers Spectrum.
Output: A cyclic peptide LeaderPeptide with amino acids taken only from
the top M elements (and ties) of the convolution of
Spectrum that fall between 57 and 200, and where the size of Leaderboard
is restricted to the top N (and ties).

Sample Input:
20
60
57 57 71 99 129 137 170 186 194 208 228 265 285 299 307 323 356 364 394
422 493
Sample Output:
99-71-137-57-72-57

Too time consuming... takes 8.5min but needs to be <5min :-1: